requirements.txt
amqp==1.4.6
ansible==1.8.4
anyjson==0.3.3
Babel==0.9.6
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
CoffeeScript==1.0.10
configobj==4.7.2
decorator==3.4.0
docker-compose==1.1.0
Django==1.7.4
django-celery==3.1.16
django-classy-tags==0.6.1
django-grappelli==2.6.3
django-oauth2-provider==0.2.6.1
djangorestframework==3.0.4
docutils==0.12
ecdsa==0.11
gunicorn==19.2.1
httplib2==0.9
importlib==1.0.3
iniparse==0.4
Jinja2==2.7.3
kombu==3.0.24
lockfile==0.10.2
lxml==3.2.1
MarkupSafe==0.23
mercurial==2.6.2
numpy==1.9.1
paramiko==1.15.2
Pillow==2.7.0
psycopg2==2.5.4
pyasn1==0.1.6
pycrypto==2.6.1
pycurl==7.19.0
PyExecJS==1.1.0
pygpgme==0.3
pyliblzma==0.5.3
python-daemon==2.0.5
python-keyczar==0.71rc0
pytz==2014.10
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
requests==2.5.1
scipy==0.15.1
shortuuid==0.4.2
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
SQLAlchemy==0.9.8
Twisted==15.0.0
urlgrabber==3.10
uWSGI==2.0.9
virtualenv==12.0.7
yum-metadata-parser==1.1.4
zope.interface==4.1.2
django-cors-headers
django-any-urlfield
django-admin-sortable
django-admin-tools
django-adminactions
django-angular
django-bootstrap
django-bootstrap-toolkit
django-bootstrap-pagination
django-uni-form
django-actions
django-extensions
django-filters
django-google-maps
django-admin-import
django-geoposition
django-gmapi
django-rules-light
i18n
funcy
